backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plurist-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=4.2",
    "django-ninja>=1.1",
    "celery[redis]>=5.3",
    "gunicorn>=21.2",
    "uvicorn[standard]>=0.30",
    "nh3>=0.2",
    "psycopg[binary]>=3.1",
    "python-decouple>=3.8",
    "httpx>=0.27",
    "authlib>=1.3",
    "requests>=2.32",
    "boto3>=1.34",
    "factory-boy>=3.3",
    "pillow>=10.0",
    "cryptography>=42.0",
    "mcp>=1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=6.0",
    "pytest-httpx>=0.30",
    "bandit[toml]>=1.8",
    "detect-secrets>=1.5",
    "django-stubs[compatible-mypy]>=5.1",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"apps/**/migrations/*.py" = ["E501"]
"apps/rendering/formats.py" = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]

[tool.coverage.run]
source = ["apps", "config"]
relative_files = true

[tool.coverage.xml]
output = "coverage.xml"

[tool.bandit]
exclude_dirs = ["backend/tests", "backend/**/migrations"]
skips = ["B101", "B105", "B106", "B110"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
pretty = true
show_error_codes = true
show_error_context = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_reexport = true
cache_dir = ".mypy_cache"
sqlite_cache = true
exclude = [
    '^tests/',
    '^apps/.*/migrations/',
]

[[tool.mypy.overrides]]
module = [
    "authlib.*",
    "boto3.*",
    "botocore.*",
    "celery.*",
    "decouple.*",
    "factory.*",
    "factory_boy.*",
    "nh3.*",
    "requests.*",
    "mcp.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "apps.generation.providers.*",
]
disallow_untyped_defs = true

[tool.django-stubs]
django_settings_module = "config.settings.test"
strict_settings = false
strict_model_abstract_attrs = true

[tool.hatch.build.targets.wheel]
packages = ["config", "apps"]
